feat: apply permission section
